pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "wove"
version = "2.0.0"
authors = [
  { name="curvedinf" },
]
description = "Inline task orchestration for Python"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = []
[project.urls]
"Homepage" = "https://github.com/curvedinf/wove"
"Bug Tracker" = "https://github.com/curvedinf/wove/issues"
[project.optional-dependencies]
dispatch = [
    "cloudpickle",
]
dev = [
    "cloudpickle",
    "pytest-asyncio",
    "ruff",
    "pytest-cov",
]
docs = [
    "sphinx>=8.0",
    "myst-parser>=4.0",
    "shibuya>=2025.9.25; python_version >= '3.10'",
]
dev-executors = [
    "cloudpickle",
    "grpcio",
    "websockets",
    "celery",
    "temporalio",
    "ray",
    "rq",
    "taskiq",
    "arq",
    "dask[distributed]",
    "kubernetes",
    "boto3",
    "pyslurm; platform_system == 'Linux'",
]
examples = [
    "httpx",
    "numpy",
    "requests",
]
